// services/model/newGalleryImageDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

export interface NewGalleryImageDto {
  /** Unique key for the image in storage */
  key: string;
  /** URL of the uploaded image */
  imageUrl: string;
  /** Order of the image for display purposes */
  order: number;
}