// services/model/updateProfileDtoFaculty.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

/**
 * User faculty
 */
export type UpdateProfileDtoFaculty = typeof UpdateProfileDtoFaculty[keyof typeof UpdateProfileDtoFaculty];


export const UpdateProfileDtoFaculty = {
  THAI: 'THAI',
  ENGLISH: 'ENGLISH',
  DTI: 'DTI',
  HRM: 'HRM',
  BBA: 'BBA',
  CIVIL: 'CIVIL',
  AVIATION: 'AVIATION',
  ENGINEERING: 'ENGINEERING',
  PSYCHOLOGY: 'PSYCHOLOGY',
  LAW: 'LAW',
  COMM_ARTS: 'COMM_ARTS',
  ARCHITECTURE: 'ARCHITECTURE',
} as const;