// services/model/updateProfileDtoGender.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

/**
 * User gender
 */
export type UpdateProfileDtoGender =
  (typeof UpdateProfileDtoGender)[keyof typeof UpdateProfileDtoGender];

export const UpdateProfileDtoGender = {
  MALE: "MALE",
  FEMALE: "FEMALE",
  OTHER: "OTHER",
} as const;