// services/model/updateProfileDtoInterestsItem.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

export type UpdateProfileDtoInterestsItem =
  (typeof UpdateProfileDtoInterestsItem)[keyof typeof UpdateProfileDtoInterestsItem];

export const UpdateProfileDtoInterestsItem = {
  BADMINTON: "BADMINTON",
  MOO_KRATA: "MOO_KRATA",
  LANGUAGE_EXCHANGE: "LANGUAGE_EXCHANGE",
  TRAVEL: "TRAVEL",
  MUSIC: "MUSIC",
  SPORTS: "SPORTS",
  GAMING: "GAMING",
  COOKING: "COOKING",
  ART: "ART",
  TECHNOLOGY: "TECHNOLOGY",
} as const;