// services/model/updateProfileDtoNationality.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

/**
 * User nationality
 */
export type UpdateProfileDtoNationality =
  (typeof UpdateProfileDtoNationality)[keyof typeof UpdateProfileDtoNationality];

export const UpdateProfileDtoNationality = {
  THAI: "THAI",
  AMERICAN: "AMERICAN",
  BRITISH: "BRITISH",
  CANADIAN: "CANADIAN",
  AUSTRALIAN: "AUSTRALIAN",
  CHINESE: "CHINESE",
  JAPANESE: "JAPANESE",
  KOREAN: "KOREAN",
  INDIAN: "INDIAN",
  VIETNAMESE: "VIETNAMESE",
  MALAYSIAN: "MALAYSIAN",
  SINGAPOREAN: "SINGAPOREAN",
  INDONESIAN: "INDONESIAN",
  PHILIPPINES: "PHILIPPINES",
  BURMESE: "BURMESE",
  LAOTIAN: "LAOTIAN",
  CAMBODIAN: "CAMBODIAN",
  BANGLADESHI: "BANGLADESHI",
  PAKISTANI: "PAKISTANI",
  AFGHAN: "AFGHAN",
  IRAQI: "IRAQI",
  SAUDI: "SAUDI",
  EMIRATI: "EMIRATI",
  QATARI: "QATARI",
  KUWAITI: "KUWAITI",
  BAHRAINI: "BAHRAINI",
  OMANI: "OMANI",
  JORDANIAN: "JORDANIAN",
  LEBANESE: "LEBANESE",
  SYRIAN: "SYRIAN",
  PALESTINIAN: "PALESTINIAN",
  ISRAELI: "ISRAELI",
  EGYPTIAN: "EGYPTIAN",
  MOROCCAN: "MOROCCAN",
  TUNISIAN: "TUNISIAN",
  ALGERIAN: "ALGERIAN",
  LIBYAN: "LIBYAN",
  SUDANESE: "SUDANESE",
  KENYAN: "KENYAN",
  NIGERIAN: "NIGERIAN",
  SOUTH_AFRICAN: "SOUTH_AFRICAN",
  ETHIOPIAN: "ETHIOPIAN",
  TANZANIAN: "TANZANIAN",
  GHANAIAN: "GHANAIAN",
  FRENCH: "FRENCH",
  GERMAN: "GERMAN",
  SPANISH: "SPANISH",
  ITALIAN: "ITALIAN",
  PORTUGUESE: "PORTUGUESE",
  DUTCH: "DUTCH",
  BELGIAN: "BELGIAN",
  SWISS: "SWISS",
  AUSTRIAN: "AUSTRIAN",
  SWEDISH: "SWEDISH",
  NORWEGIAN: "NORWEGIAN",
  DANISH: "DANISH",
  FINNISH: "FINNISH",
  POLISH: "POLISH",
  CZECH: "CZECH",
  HUNGARIAN: "HUNGARIAN",
  ROMANIAN: "ROMANIAN",
  BULGARIAN: "BULGARIAN",
  GREEK: "GREEK",
  TURKISH: "TURKISH",
  RUSSIAN: "RUSSIAN",
  UKRAINIAN: "UKRAINIAN",
  BELARUSIAN: "BELARUSIAN",
  SERBIAN: "SERBIAN",
  CROATIAN: "CROATIAN",
  BOSNIAN: "BOSNIAN",
  SLOVENIAN: "SLOVENIAN",
  LITHUANIAN: "LITHUANIAN",
  LATVIAN: "LATVIAN",
  ESTONIAN: "ESTONIAN",
  MEXICAN: "MEXICAN",
  BRAZILIAN: "BRAZILIAN",
  ARGENTINE: "ARGENTINE",
  COLOMBIAN: "COLOMBIAN",
  CHILEAN: "CHILEAN",
  PERUVIAN: "PERUVIAN",
  VENEZUELAN: "VENEZUELAN",
  ECUADORIAN: "ECUADORIAN",
  BOLIVIAN: "BOLIVIAN",
  PARAGUAY: "PARAGUAY",
  URUGUAY: "URUGUAY",
  GUATEMALAN: "GUATEMALAN",
  HONDURAS: "HONDURAS",
  COSTA_RICAN: "COSTA_RICAN",
  PANAMANIAN: "PANAMANIAN",
  CUBAN: "CUBAN",
  DOMINICAN: "DOMINICAN",
  JAMAICAN: "JAMAICAN",
  TRINIDADIAN: "TRINIDADIAN",
  BAHAMANIAN: "BAHAMANIAN",
  BARBADIAN: "BARBADIAN",
  NEW_ZEALANDER: "NEW_ZEALANDER",
  FIJIAN: "FIJIAN",
  THAI_STATELESS: "THAI_STATELESS",
  OTHER: "OTHER",
} as const;