// services/model/updateProfileDtoPreferredGender.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * KBU Connect API
 * Exclusive dating platform for KBU students
 * OpenAPI spec version: 1.0
 */

/**
 * Preferred gender for matching
 */
export type UpdateProfileDtoPreferredGender =
  (typeof UpdateProfileDtoPreferredGender)[keyof typeof UpdateProfileDtoPreferredGender];

export const UpdateProfileDtoPreferredGender = {
  MALE: "MALE",
  FEMALE: "FEMALE",
  ALL: "ALL",
} as const;